build: Do not use headers on library building
Headers are not necessary to be passed to the library compilation
function because the compiler will find them. On the other hand
they are necessary for the proper GIR generation.

This patch splits headers and sources, uses only sources for the
library building and uses both for GIR generation. It also allows
getting both separately.

https://bugzilla.gnome.org/show_bug.cgi?id=792948