mixer-control: Remove unnecessary instance checks
gvc_mixer_control_get_vol_max_norm() and gvc_mixer_control_get_vol_max_amplified()
do not access any of the instance variables. In GNOME Settings, we even
pass NULL to these functions. So checking the instance is not necessary.

In fact, this is causing instance check failures in GNOME Settings, and
the entire Sound panel is broken since those functions are returning zero.

The reason that only the checks are being dropped, and not the parameter
entirely, is to preserve API compatibility.